src/requirements.txt
alembic==1.8.1
aniso8601==9.0.1
attrs==23.1.0
bcrypt==3.2.2
cffi==1.15.1
click==8.1.6
cryptography==41.0.3
Flask==2.1.3
Flask-Cors==3.0.10
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
flask-restx==0.5.1
Flask-SQLAlchemy==2.5.1
gunicorn==21.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.9.1
loguru==0.6.0
Mako==1.2.4
MarkupSafe==2.1.3
marshmallow==3.17.1
marshmallow-sqlalchemy==0.28.2
packaging==21.3
pkgutil_resolve_name==1.3.10
psycopg2-binary==2.9.7
pycparser==2.21
PyJWT==2.4.0
PyNaCl==1.5.0
pyparsing==3.1.1
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.7.1
six==1.16.0
SQLAlchemy==1.4.49
Werkzeug==2.1.2
